open browser and try to login
