+++ PyViCare/PyViCareOAuthManager.py
@@ -1,14 +1,10 @@
-from PyViCare import Feature
-from PyViCare.PyViCareUtils import PyViCareInvalidCredentialsError, PyViCareRateLimitError
-from abc import abstractclassmethod
-from oauthlib.oauth2 import TokenExpiredError
+from PyViCare.PyViCareAbstractOAuthManager import AbstractViCareOAuthManager
+from PyViCare.PyViCareUtils import PyViCareInvalidCredentialsError
 import requests
 import re
 import pickle
-import threading
 import os
 import pkce
-from http.server import BaseHTTPRequestHandler, HTTPServer
 from pickle import UnpicklingError
 from requests_oauthlib import OAuth2Session
 import logging
@@ -21,115 +17,6 @@ AUTHORIZE_URL = 'https://iam.viessmann.com/idp/v2/authorize'
 TOKEN_URL = 'https://iam.viessmann.com/idp/v2/token'
 REDIRECT_URI = "vicare://oauth-callback/everest"
 VIESSMANN_SCOPE = ["IoT User"]
-API_BASE_URL = 'https://api.viessmann.com/iot/v1'
-
-
-class AbstractViCareOAuthManager:
-    def __init__(self):
-        self.oauth = None
-
-    @abstractclassmethod
-    def renewToken(self):
-        return
-
-    def get(self, url):
-        try:
-            logger.debug(self.oauth)
-            response = self.oauth.get(f"{API_BASE_URL}{url}").json()
-            logger.debug("Response to get request: "+str(response))
-            self.handleExpiredToken(response)
-            self.handleRateLimit(response)
-            return response
-        except TokenExpiredError:
-            self.renewToken()
-            return self.get(url)
-
-    def handleExpiredToken(self, response):
-        if("error" in response and response["error"] == "EXPIRED TOKEN"):
-            raise TokenExpiredError(response)
-
-    def handleRateLimit(self, response):
-        if not Feature.raise_exception_on_rate_limit:
-            return
-
-        if("statusCode" in response and response["statusCode"] == 429):
-            raise PyViCareRateLimitError(response)
-
-    """POST URL using OAuth session. Automatically renew the token if needed
-    Parameters
-    ----------
-    url : str
-        URL to get
-    data : str
-        Data to post
-
-    Returns
-    -------
-    result: json
-        json representation of the answer
-    """
-
-    def post(self, url, data):
-        headers = {"Content-Type": "application/json",
-                   "Accept": "application/vnd.siren+json"}
-        try:
-            response = self.oauth.post(
-                f"{API_BASE_URL}{url}", data, headers=headers).json()
-            self.handleExpiredToken(response)
-            self.handleRateLimit(response)
-            return response
-        except TokenExpiredError:
-            self.renewToken()
-            return self.post(url, data)
-
-
-class ViCareHomeAssistantOAuthManager(AbstractViCareOAuthManager):
-    def __init__(self, oauth):
-        self.oauth = oauth
-
-    def renewToken(self):
-        # todo: move this class to Home Assistant and implement the oauth flow
-        return super().renewToken()
-
-class ViCareBrowserOAuthManager(AbstractViCareOAuthManager):
-    class Serv(BaseHTTPRequestHandler):
-        def __init__(self, callback, *args):
-            self.callback = callback
-            BaseHTTPRequestHandler.__init__(self, *args)
-
-        def do_GET(self):
-            self.callback(self.path)
-            self.send_response(200)
-            self.send_header("Content-type", "text/plain")
-            self.end_headers()
-            self.wfile.write("Success. You can close this browser window now.".encode("utf-8"))
-            
-    def __init__(self, client_id, token_file):
-        self.token_file = token_file
-        self.client_id = client_id
-        self.abc = self.__createNewSession()
-
-    def __createNewSession(self):
-        server = None
-        code = None
-        def callback(path):
-            nonlocal code
-            nonlocal server
-            match = re.match(r"(?P<uri>.+?)\?code=(?P<code>[^&]+)", path)
-            code = match.group('code')
-            threading.Thread(target=server.shutdown, daemon=True).start()
-
-        def handlerWithCallbackWrapper(*args):
-            ViCareBrowserOAuthManager.Serv(callback, *args)
-
-        server = HTTPServer(('localhost',5125), handlerWithCallbackWrapper)
-        server.serve_forever()
-        print(f"Code: {code}")
-
-    def renewToken(self):
-        # todo: move this class to Home Assistant and implement the oauth flow
-        return super().renewToken()
-
 
 class ViCareOAuthManager(AbstractViCareOAuthManager):
     def __init__(self, username, password, client_id, token_file):
